Feature/UI improvements (#27)
* initial commit of frontend v2

* new templates for news, articles and categories pages

* api interaction with backend

* categories interactivity

* news interactivity

* added article timestamp. added new pages

* category and subcategory changes. Improvements in error handling

* models refactoring

* articles pages implementation

* some articles-related components improvements. category components created. changed favicon

* header logo change. news-related components updates

* implemented signing up and email confirmation

* some authorization improvements

* authorization related components

* auth improvements

* some style improvements. initial account page

* packages upgrade

* news and article page changes

* auth improvements on backend side

* auth improvements

* initial profile and user articles pages

* conditionally generate token on sign in

* initial welcome page. some style changes in other pages

* account page

* fix substring error

* admin page. other pages improvements

* claims usage improvements

* some website parts restrictions for non-authorized and non-administrator users

* restrictions to some endpoints. security improvements

* update database script improvements

* handle exceptions on connection to rabbitmq

* handle reconnection to rabbitmq on error

* loading spinner while guards execution. initial localization

* localization improvements

* localization improvements

* localization of articles page

* articles, categories and news pages localization

* account page localization

* admin page localization

* fix getting article by id

* fix translations

+++ ScienceArchive.Server/ScienceArchive.Application/UseCases/ArticleUseCases/DeleteArticleUseCase.cs
@@ -2,7 +2,11 @@ using ScienceArchive.Application.Abstractions.Persistence;
 using ScienceArchive.Application.Dtos.Article.Request;
 using ScienceArchive.Application.Dtos.Article.Response;
 using ScienceArchive.Application.Interfaces;
+using ScienceArchive.Core.Domain.Aggregates.Article;
+using ScienceArchive.Core.Domain.Aggregates.Article.Enums;
 using ScienceArchive.Core.Domain.Aggregates.Article.ValueObjects;
+using ScienceArchive.Core.Domain.Aggregates.User.ValueObjects;
+using ScienceArchive.Core.Exceptions;
 
 namespace ScienceArchive.Application.UseCases.ArticleUseCases;
 
@@ -18,6 +22,19 @@ internal class DeleteArticleUseCase : IUseCase<DeleteArticleRequestDto, DeleteAr
     public async Task<DeleteArticleResponseDto> Handle(DeleteArticleRequestDto request, CancellationToken cancellationToken)
     {
         var articleId = ArticleId.CreateFromString(request.Id);
+        var userId = UserId.CreateFromString(request.UserId);
+        var article = await _dbUnitOfWork.ArticleRepository.GetById(articleId);
+
+        if (article is null)
+        {
+            throw new EntityNotFoundException(nameof(Article));
+        }
+
+        if (!article.Authors.Any(a => a.Role == ArticleAuthorRole.Owner && a.UserId.Equals(userId)))
+        {
+            throw new IncorrectArticleCreatorException();
+        }
+        
         var deletedArticleId = await _dbUnitOfWork.ArticleRepository.Delete(articleId);
         
         return new DeleteArticleResponseDto(deletedArticleId.ToString());
