+++ ScienceArchive.Server/ScienceArchive.Web.Api/Controllers/ArticleController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using ScienceArchive.Application.Dtos.Article.Request;
 using ScienceArchive.Application.Interfaces.Services;
+using ScienceArchive.Web.Api.Auth;
 using ScienceArchive.Web.Api.Responses;
 using ScienceArchive.Web.Api.Utils;
 
@@ -17,18 +19,19 @@ public class ArticleController : ControllerBase
     }
     
     [HttpGet("by-category/{categoryId}")]
-    public async Task<Response> GetByCategoryId(string categoryId)
+    public async Task<Response> GetVerifiedByCategoryId(string categoryId)
     {
         if (string.IsNullOrWhiteSpace(categoryId))
         {
             throw new BadHttpRequestException("Category ID was not presented");
         }
         
-        var dto = new GetArticlesByCategoryIdRequestDto(categoryId);
-        var result = await _articleService.GetArticlesByCategoryId(dto);
+        var dto = new GetVerifiedArticlesByCategoryIdRequestDto(categoryId);
+        var result = await _articleService.GetVerifiedArticlesByCategoryId(dto);
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpGet("my-articles")]
     public async Task<Response> GetUserArticles()
     {
@@ -63,12 +66,16 @@ public class ArticleController : ControllerBase
         {
             throw new BadHttpRequestException("ID was not presented");
         }
-        
-        var dto = new GetArticleByIdRequestDto(id);
+
+        var userId = HttpContext.GetUserIdFromToken();
+
+        var requiredClaims = new List<string> { AuthClaims.ViewDeclinedArticles, AuthClaims.ViewNotVerifiedArticles };
+        var dto = new GetArticleByIdRequestDto(id, userId, requiredClaims);
         var result = await _articleService.GetArticleById(dto);
         return new SuccessResponse(result);
     }
 
+    [AuthorizeClaims(AuthClaims.ViewNotVerifiedArticles, AuthClaims.ViewDeclinedArticles)]
     [HttpGet("all")]
     public async Task<Response> GetAll()
     {
@@ -87,6 +94,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpPost("create")]
     public async Task<Response> Create([FromBody] CreateArticleRequestDto? dto)
     {
@@ -95,10 +103,19 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var result = await _articleService.CreateArticle(dto);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+
+        var populatedDto = dto with { UserId = userId };
+        var result = await _articleService.CreateArticle(populatedDto);
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpPost("update")]
     public async Task<Response> Update([FromBody] UpdateArticleRequestDto? dto)
     {
@@ -107,10 +124,19 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var result = await _articleService.UpdateArticle(dto);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+
+        var populatedDto = dto with { UserId = userId};
+        var result = await _articleService.UpdateArticle(populatedDto);
         return new SuccessResponse(result);
     }
 
+    [AuthorizeClaims(AuthClaims.ApproveArticles)]
     [HttpPost("approve")]
     public async Task<Response> Approve([FromBody] ApproveArticleRequestDto? dto)
     {
@@ -123,7 +149,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
-    
+    [AuthorizeClaims(AuthClaims.DeclineArticles)]
     [HttpPost("decline")]
     public async Task<Response> Decline([FromBody] DeclineArticleRequestDto? dto)
     {
@@ -136,6 +162,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<Response> Delete(string? id)
     {
@@ -144,7 +171,14 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var dto = new DeleteArticleRequestDto(id);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+        
+        var dto = new DeleteArticleRequestDto(id, userId);
         var result = await _articleService.DeleteArticle(dto);
         return new SuccessResponse(result);
     }
